refactors utils and adds some caching logic

+++ api/groa_ds_api/utils.py
@@ -16,35 +16,32 @@ class MovieUtility(object):
     def __init__(self, model_path):
         """ Initialize model with name of .model file """
         self.model_path = model_path
-        self.model = self._load_model()
-        self.connection = psycopg2.connect(
+        self.model = self.__load_model()
+        self.connection = self.__get_connection()
+        self.id_book = self.__get_id_book()
+    
+    # ------- Start Private Methods -------
+    def __get_connection(self):
+        return psycopg2.connect(
             database  =  os.getenv('DB_NAME'),
             user      =  os.getenv('DB_USER'),
             password  =  os.getenv('DB_PASSWORD'),
             host      =  os.getenv('HOST'),
             port      =  os.getenv('PORT')
         )
-        self.id_book = self._get_id_book()
     
-    # ------- Start Helper Methods -------
-    def _get_cursor(self):
+    def __get_cursor(self):
         """ Grabs cursor from self.connection """
         try:
             cursor = self.conneciton.cursor()
             return cursor 
         except:
-            self.connection = psycopg2.connect(
-                database  =  os.getenv('DB_NAME'),
-                user      =  os.getenv('DB_USER'),
-                password  =  os.getenv('DB_PASSWORD'),
-                host      =  os.getenv('HOST'),
-                port      =  os.getenv('PORT')
-            )
+            self.connection = self.__get_connection()
             return self.connection.cursor()
 
-    def _get_id_book(self):
+    def __get_id_book(self):
         """ Gets movie data from database to merge with recommendations """
-        self.cursor_dog = self._get_cursor()
+        self.cursor_dog = self.__get_cursor()
         query = "SELECT movie_id, primary_title, start_year, genres, poster_url FROM movies;"
         self.cursor_dog.execute(query)
         movie_sql= self.cursor_dog.fetchall()
@@ -52,7 +49,7 @@ class MovieUtility(object):
         self.cursor_dog.close()
         return id_book
 
-    def _load_model(self):
+    def __load_model(self):
         """ Get the model object for this instance, loading it if it's not already loaded """
         w2v_model = gensim.models.Word2Vec.load(self.model_path)
         # Keep only the normalized vectors.
@@ -61,12 +58,18 @@ class MovieUtility(object):
         self.model = w2v_model
         return self.model
 
-    def _get_info(self, recs):
+    def __get_info(self, recs):
         """ Merging recs with id_book to get movie info """
         return pd.merge(recs, self.id_book, how='left', on='movie_id')
     
-    def _get_JSON(self, rec_df):
-        """ Turn predictions into JSON """
+    def __get_JSON(self, rec_df):
+        """ 
+        Turn predictions into JSON
+        Callers:
+            - get_recommendations
+            - get_similar_movies
+            - get_movie_list
+        """
         names = rec_df.columns
         rec_json = []
 
@@ -83,12 +86,74 @@ class MovieUtility(object):
                     })
 
         return rec_json
+    
+    def __prep_data(self, ratings_df, watched_df=None, watchlist_df=None, good_threshold=4, bad_threshold=3):
+        """Converts dataframes of exported Letterboxd data to lists of movie_ids.
 
-    def _predict(self, input, bad_movies=[], hist_list=[], val_list=[],
+        Parameters
+        ----------
+        ratings_df : pd dataframe
+            user ratings.
+
+        watched_df : pd dataframe
+            user watch history.
+
+        watchlist_df : pd dataframe
+            list of movies the user wants to watch.
+            Used in val_list for scoring the model's performance.
+
+        good_threshold : int
+            Minimum star rating (5pt scale) for a movie to be considered "enjoyed" by the user.
+
+        bad_threshold : int
+            Maximum star rating (5pt scale) for a movie to be considered "disliked" by the user.
+
+
+        Returns
+        -------
+        tuple of lists of ids.
+            (good_list, bad_list, hist_list, val_list, ratings_dict)
+            point scale seems to be 5
+            good_list is ratings greater than good_threashold
+            bad_list is ratings below bad_threshold
+            neutral_list is movies not in good or bad list
+            val_list is just movies in watched_df
+            hist_list seems to be movies not in good or bad too?
+            ratings_dict all ids mapped to their ratings
+        """
+        try:
+            # split according to user rating
+            good_df = ratings_df[ratings_df['rating'] >= good_threshold]
+            bad_df = ratings_df[ratings_df['rating'] <= bad_threshold]
+            neutral_df = ratings_df[(ratings_df['rating'] > bad_threshold) & (ratings_df['rating'] < good_threshold)]
+            
+            # convert dataframes to lists
+            good_list = good_df['movie_id'].to_list()
+            bad_list = bad_df['movie_id'].to_list()
+            neutral_list = neutral_df['movie_id'].to_list()
+
+        except Exception as e:
+            print("Error making good, bad and neutral list")
+            raise Exception(e)
+
+        ratings_dict = pd.Series(ratings_df['rating'].values,index=ratings_df['movie_id']).to_dict()
+
+        if watched_df is not None:
+            # Construct list of watched movies that aren't rated "good" or "bad"
+            hist_list = ratings_df[~ratings_df['movie_id'].isin(good_list+bad_list)]['movie_id'].to_list()
+        else: hist_list = neutral_list
+
+        if watchlist_df is not None:
+            # gets list of movies user wants to watch for validation
+            val_list = watchlist_df['movie_id'].tolist()
+        else: val_list = []
+
+        return (good_list, bad_list, hist_list, val_list, ratings_dict)
+
+    def __predict(self, input, bad_movies=[], hist_list=[], val_list=[],
                 ratings_dict = {}, checked_list=[], rejected_list=[],
                 n=50, harshness=1):
-        """Returns a list of recommendations and useful metadata, given a pretrained
-        word2vec model and a list of movies.
+        """Returns a list of recommendations, given a list of movies.
 
         Parameters
         ----------
@@ -104,9 +169,6 @@ class MovieUtility(object):
 
             val_list : iterable
                 List of movies the user has already indicated interest in.
-                Example: https://letterboxd.com/tabula_rasta/watchlist/
-                People really load these up over the years, and so they make for
-                the best validation set we can ask for with current resources.
 
             ratings_dict : dictionary
                 Dictionary of movie_id keys, user rating values.
@@ -126,24 +188,10 @@ class MovieUtility(object):
                     1 - most strongly account for disliked movies.
                     3 - divide "disliked movies" vector by 3.
 
-            rec_movies : boolean
-                If False, doesn't return movie recommendations (used for scoring).
-
-            show_vibes : boolean
-                If True, prints out the dupes as a feature.
-                These movies are closest to the user's taste vector,
-                indicating some combination of importance and popularity.
-
-            scoring : boolean
-                If True, prints out the validation score.
-
-            return_scores : boolean
-                If True, skips printing out
-
         Returns
         -------
         A list of tuples
-            (Title, Year, IMDb URL, Average Rating, Number of Votes, Similarity score)
+            (Movie ID, Similarity score)
         """
 
         clf = self.model
@@ -205,20 +253,21 @@ class MovieUtility(object):
         recs = _remove_dupes(recs, input, bad_movies, hist_list, checked_list + rejected_list)
         return recs
 
-    def _get_list_preview(self, data):
-        """ Turns list preview sql into an object """
+    def __get_list_preview(self, data):
+        """ 
+        Turns list preview sql into an object
+        Callers:
+            - get_user_lists
+            - get_all_lists 
+        """
         return {
             "list_id": data[0], 
             "name": data[1],
             "private": data[2]
         }
     
-    def _run_query(self, query, params, commit=False, fetch="one"):
-        try:
-            self.cursor_dog = self._get_cursor()
-        except Exception as e:
-            print("Connection problem chief!\n")
-            print(e)
+    def __run_query(self, query, params, commit=False, fetch="one"):
+        self.cursor_dog = self.__get_cursor()
         if params is None:
             self.cursor_dog.execute(query)
         else:
@@ -232,14 +281,14 @@ class MovieUtility(object):
             self.connection.commit()
         self.cursor_dog.close()
         return result
-    # ------- End Helper Methods -------
+    # ------- End Private Methods -------
 
     # ------- Start Public Methods -------
     def create_movie_list(self, payload):
         """ Creates a MovieList """
         query = """INSERT INTO movie_lists
         (user_id, name, private) VALUES (%s, %s, %s) RETURNING list_id;"""
-        list_id = self._run_query(
+        list_id = self.__run_query(
             query, 
             (payload.user_id, payload.name, payload.private),
             commit=True)
@@ -254,7 +303,7 @@ class MovieUtility(object):
         query = """SELECT l.movie_id, m.primary_title, m.start_year, m.genres, m.poster_url 
         FROM list_movies AS l LEFT JOIN movies AS m ON l.movie_id = m.movie_id
         WHERE l.list_id = %s;"""
-        list_sql = self._run_query(
+        list_sql = self.__run_query(
             query, 
             (list_id,),
             fetch="all")
@@ -273,40 +322,40 @@ class MovieUtility(object):
                     "genres": movie[3].split(","),
                     "poster_url": movie[4]
                 })
-            w2v_preds = self._predict(movie_ids)
+            w2v_preds = self.__predict(movie_ids)
             df_w2v = pd.DataFrame(w2v_preds, columns=['movie_id', 'score'])
             # get movie info using movie_id
-            rec_data = self._get_info(df_w2v)
+            rec_data = self.__get_info(df_w2v)
             rec_data = rec_data.fillna("None")
-            rec_json = self._get_JSON(rec_data)
+            rec_json = self.__get_JSON(rec_data)
             list_json["recs"] = rec_json
         return list_json
     
     def get_user_lists(self, user_id):
         """ Get user's MovieLists """
         query = "SELECT list_id, name, private FROM movie_lists WHERE user_id = %s;"
-        user_lists = self._run_query(
+        user_lists = self.__run_query(
             query, 
             (user_id,),
             fetch="all")
-        user_lists_json = [self._get_list_preview(elem) for elem in user_lists]
+        user_lists_json = [self.__get_list_preview(elem) for elem in user_lists]
         return user_lists_json
     
     def get_all_lists(self):
         """ Get all MovieLists """
         query = "SELECT list_id, name, private FROM movie_lists WHERE private=FALSE;"
-        lists = self._run_query(
+        lists = self.__run_query(
             query, 
             None,
             fetch="all")
-        lists_json = [self._get_list_preview(elem) for elem in lists]
+        lists_json = [self.__get_list_preview(elem) for elem in lists]
         return lists_json
     
     def add_to_movie_list(self, list_id, movie_id):
         """ Add movie to a MovieList """
         query = """INSERT INTO list_movies
         (list_id, movie_id) VALUES (%s, %s);"""
-        self._run_query(
+        self.__run_query(
             query, 
             (list_id, movie_id),
             commit=True,
@@ -316,7 +365,7 @@ class MovieUtility(object):
     def remove_from_movie_list(self, list_id, movie_id):
         """ Remove movie from a MovieList """
         query = "DELETE FROM list_movies WHERE list_id = %s AND movie_id = %s;"
-        self._run_query(
+        self.__run_query(
             query, 
             (list_id, movie_id),
             commit=True,
@@ -326,13 +375,13 @@ class MovieUtility(object):
 
     def delete_movie_list(self, list_id):
         """ Delete a MovieList """
-        query = "DELETE FROM movie_lists WHERE list_id = %s;"
-        self._run_query(
+        query = "DELETE FROM movie_lists WHERE list_id = %s RETURNING user_id, private;"
+        result = self.__run_query(
             query, 
             (list_id,),
             commit=True,
-            fetch="none")
-        return "Success"
+            fetch="all")[0]
+        return result
 
     def get_most_similar_title(self, id, id_list):
         """ Get the title of the most similar movie to id from id_list """
@@ -347,12 +396,7 @@ class MovieUtility(object):
     
     def get_service_providers(self, movie_id):
         """ Get the service providers of a given movie_id """
-        try:
-            self.cursor_dog = self._get_cursor()
-        except Exception as e:
-            print("Connection problem chief!\n")
-            print(e)
-        
+        self.cursor_dog = self.__get_cursor()
         query = """
         SELECT m.provider_id, p.name, p.logo_url, m.provider_movie_url, 
         m.presentation_type, m.monetization_type
@@ -386,9 +430,9 @@ class MovieUtility(object):
         # could check if id is in vocab
         m_vec = clf[movie_id]
         movies_df = pd.DataFrame(clf.similar_by_vector(m_vec, topn=n+1)[1:], columns=['movie_id', 'score'])
-        result_df = self._get_info(movies_df)
+        result_df = self.__get_info(movies_df)
         return {
-            "data": self._get_JSON(result_df)
+            "data": self.__get_JSON(result_df)
             }
     
     def get_recommendations(self, payload, background_tasker):
@@ -401,11 +445,7 @@ class MovieUtility(object):
         harshness = payload.harshness
 
         # create cursor
-        try:
-            self.cursor_dog = self._get_cursor()
-        except Exception as e:
-            print("Connection problem chief!\n")
-            print(e)
+        self.cursor_dog = self.__get_cursor()
 
         # Check if user has ratings data 
         query = "SELECT date, movie_id, rating FROM user_ratings WHERE user_id=%s;"
@@ -433,16 +473,16 @@ class MovieUtility(object):
         willnotwatchlist_df = pd.DataFrame(willnotwatch_sql, columns = ['date', 'movie_id'])
 
         # Prepare data
-        good_list, bad_list, hist_list, val_list, ratings_dict = prep_data(
+        good_list, bad_list, hist_list, val_list, ratings_dict = self.__prep_data(
             ratings, watched, watchlist, good_threshold=good_threshold, bad_threshold=bad_threshold
             )
 
         # Run prediction with parameters then wrangle output
-        w2v_preds = self._predict(good_list, bad_list, hist_list, val_list, ratings_dict, harshness=harshness, n=n)
+        w2v_preds = self.__predict(good_list, bad_list, hist_list, val_list, ratings_dict, harshness=harshness, n=n)
         df_w2v = pd.DataFrame(w2v_preds, columns=['movie_id', 'score'])
 
         # get movie info using movie_id
-        rec_data = self._get_info(df_w2v)
+        rec_data = self.__get_info(df_w2v)
         rec_data = rec_data.fillna("None")
 
         def _commit_to_database(model_recs, user_id, num_recs, good, bad, harsh): 
@@ -472,7 +512,7 @@ class MovieUtility(object):
             self.connection.commit()
             self.cursor_dog.close()
 
-        rec_json = self._get_JSON(rec_data)
+        rec_json = self.__get_JSON(rec_data)
 
         # add background task to commit recs to DB
         background_tasker.add_task(
@@ -483,67 +523,3 @@ class MovieUtility(object):
                 "data": rec_json
             }
     # ------- End Public Methods -------
-
-
-def prep_data(ratings_df, watched_df=None, watchlist_df=None, good_threshold=4, bad_threshold=3):
-    """Converts dataframes of exported Letterboxd data to lists of movie_ids.
-
-    Parameters
-    ----------
-    ratings_df : pd dataframe
-        user ratings.
-
-    watched_df : pd dataframe
-        user watch history.
-
-    watchlist_df : pd dataframe
-        list of movies the user wants to watch.
-        Used in val_list for scoring the model's performance.
-
-    good_threshold : int
-        Minimum star rating (5pt scale) for a movie to be considered "enjoyed" by the user.
-
-    bad_threshold : int
-        Maximum star rating (5pt scale) for a movie to be considered "disliked" by the user.
-
-
-    Returns
-    -------
-    tuple of lists of ids.
-        (good_list, bad_list, hist_list, val_list, ratings_dict)
-        point scale seems to be 5
-        good_list is ratings greater than good_threashold
-        bad_list is ratings below bad_threshold
-        neutral_list is movies not in good or bad list
-        val_list is just movies in watched_df
-        hist_list seems to be movies not in good or bad too?
-        ratings_dict all ids mapped to their ratings
-    """
-    try:
-        # split according to user rating
-        good_df = ratings_df[ratings_df['rating'] >= good_threshold]
-        bad_df = ratings_df[ratings_df['rating'] <= bad_threshold]
-        neutral_df = ratings_df[(ratings_df['rating'] > bad_threshold) & (ratings_df['rating'] < good_threshold)]
-        
-        # convert dataframes to lists
-        good_list = good_df['movie_id'].to_list()
-        bad_list = bad_df['movie_id'].to_list()
-        neutral_list = neutral_df['movie_id'].to_list()
-
-    except Exception as e:
-        print("Error making good, bad and neutral list")
-        raise Exception(e)
-
-    ratings_dict = pd.Series(ratings_df['rating'].values,index=ratings_df['movie_id']).to_dict()
-
-    if watched_df is not None:
-        # Construct list of watched movies that aren't rated "good" or "bad"
-        hist_list = ratings_df[~ratings_df['movie_id'].isin(good_list+bad_list)]['movie_id'].to_list()
-    else: hist_list = neutral_list
-
-    if watchlist_df is not None:
-        # gets list of movies user wants to watch for validation
-        val_list = watchlist_df['movie_id'].tolist()
-    else: val_list = []
-
-    return (good_list, bad_list, hist_list, val_list, ratings_dict)
